backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
motor==3.7.0
pymongo==4.11.1
redis==5.2.1
python-dotenv==1.0.1
pydantic-settings==2.8.1
python-multipart==0.0.20
cloudinary==1.44.0
qrcode[pil]==8.0
jinja2==3.1.5
twilio==9.4.4
email-validator==2.2.0
